DUI: Use DocumentsProvider API for handling custom icons [1/2]
All images requested for DUI features will be copied and kept
in a folder in external storage. We can not rely on third party
content providers to deliver our icons to SystemUI. Either the provider
gets us a good icon and we copy it successfully or we toast an error
and let SystemUI know the request was canceled.

In time, we may wish to consider a "IconProvider" type of solution

AICPfy

Change-Id: I395f5e617ef0f31b3b19de2dd8fd224ad398e214
Signed-off-by: scott <scott.hart.bti@gmail.com>

Conflicts:
	src/com/android/settings/cyanogenmod/IconPickerGallery.java